Refactor DataCache implementation w/ more modular approach
